remove complex answers from set operations and place in new class which combines many operations
