+++ LINQ.Exercises/LINQ.Exercises/SetOperations.cs
@@ -18,7 +18,6 @@ namespace LINQ.Exercises
     [TestClass]
     public class SetOperations
     {
-
         // I want to see an enumerable which contains only the unique numbers in the
         // above array
 
@@ -32,7 +31,6 @@ namespace LINQ.Exercises
             Assert.IsTrue(result.SequenceEqual(new int[] { 2, 3, 5, 4, 6, 8, 7, 9, 34, 67 }));
         }
 
-
         // get the unique numbers
         // from both arrays combined
         // in ascending order from 1...9
@@ -42,7 +40,7 @@ namespace LINQ.Exercises
             int[] numbersA = { 0, 2, 4, 5, 6, 8, 9 };
             int[] numbersB = { 1, 3, 5, 7, 8 };
 
-            IEnumerable<int> result = numbersA; 
+            IEnumerable<int> result = numbersA;
 
             Assert.IsTrue(result.SequenceEqual(new int[] { 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 }));
         }
@@ -53,49 +51,18 @@ namespace LINQ.Exercises
             int[] numbersA = { 0, 2, 4, 5, 6, 8, 9 };
             int[] numbersB = { 1, 3, 5, 7, 8 };
 
-            IEnumerable<int> result = numbersA; 
+            IEnumerable<int> result = numbersA;
 
             Assert.IsTrue(result.SequenceEqual(new int[] { 5, 8 }));
         }
-
-
-        // we have a list of people
-        // we have their first names
-        // i need you to find out which letters are common
-        // to all the first names
-
-        // Hint: try to use set operations.
-        // There are many ways to solve this.
-        [TestMethod]
-        public void GetCharactersCommonToEveryonesFirstNamesUsingSetElements_ReturnCharEnumerable()
-        {
-            
-            List<char> commonCharacters = new List<char>(); // please edit/complete so that the test passes
-
-            Assert.IsTrue(commonCharacters.OrderBy(x => x).SequenceEqual(new char[] { 'a', 'i', 'J' }.OrderBy(x => x)));
-        }
-
-        // Bonus Question
-        // we have a list of people
-        // we have their first names
-        // i need you to find out which letters are common
-        // to all the first names names
-        // But you are not allowed to use set operations.
-        [TestMethod]
-        public void GetCharactersCommonToEveryonesFirstNamesNotUsingSetOperations_ReturnCharEnumerable()
-        {
-            IEnumerable<char> result = new List<char>();
-
-            Assert.IsTrue(result.OrderBy(x => x).SequenceEqual(new char[] { 'a', 'i', 'J' }.OrderBy(x => x)));
-        }
-
+               
         [TestMethod]
         public void GetNumbersInFirstArrayThatAreNotAlsoInSecondArray_ReturnIenumerableInt()
         {
             int[] numbersA = { 0, 2, 4, 5, 6, 8, 9 };
             int[] numbersB = { 1, 3, 5, 7, 8 };
 
-            IEnumerable<int> result = numbersA; 
+            IEnumerable<int> result = numbersA;
 
             Assert.IsTrue(result.OrderBy(x => x).SequenceEqual(new int[] { 0, 2, 4, 6, 9 }.OrderBy(x => x)));
         }
